Improve original files backup and restore strategy of install AMP
Original files were backed up and editing was done on the original
files. But if the install process was done multiple times, some renamed
files could have gone missing.
Now the strategy is to rename the original files to a backup name and
then copy them to their original path. Editing is done on the restored
original files (not files that might have been already edited).

Also, the default php fpm pool file is now backed up in original `/etc`
directory and not in the `/root` directory. There is no need to move the
backup file to `/root` directory because files with a `.orig` extension
are not processed by php fpm.