Add a script to install phpMyAdmin from source
When the debian package of phpMyAdmin is not avialable, a script to install
phpMyAdmin from source is needed.

The new script installs and configures phpMyAdmin. But it also adds an http
authentication password to secure the web application.

A script to uninstall phpMyAdmin is also provided.

The script to reset phpMyAdmin http authentication password has also been
completely updated.

Finaly, the existing script to install phpMyAdmin from package was partialy
updated to reflect some of the improvements brought by script that installs
from source.

Note that the script to create a new database needs an important update.
See it's "@fixme" comment for details.