Disable problématc scripts
The old firewall install script was removed because it was unable to
handle custom ssh port.

The new security config script (fail2ban, firewall, ssh) is currently
disabled because it is still work in progress.

The script which installs phpMyAdmin from debien package has been
temporarily disabled because the package is currently unavailable.
It also needs updates to make the code more similar to the script
which installs phpMyAdmin from source.

The documentation has been updated to reflect all these changes.