Fix php fpm session security issue
For a given client but different php fpm pools, php fpm creates session data
file with the same name.
If a different session directories for different pools is not specified,
two different pools could share the same session data, which is obviously
a major security issue.
This fix specifies a dedicated session directory for each php fpm pool.
A specific pool is also specified for phpmyadmin.

This fix also restricts the open_basedir parameter to the project directory
for each php fpm pool.

The removal of the session directories is also taken care of.