Improve documentation of mkwebsite.sh
The documentation was using the word LOCAL_DOMAIN though the domain could
have been a global domain name.
The word has been repaced by DOMAIN, which can reference local domain as
well as global domain name.

The documentation did not mention the existence or the use of the
"template-subdir.sh" or "template-subdir-symfony.sh".
Now the documentation explains the use of these templates and give a
examples of commands to use them.

When using the "template-subdir*.sh" templates, the DOMAIN parameter is
isgnored.
This is information is now expicit for the user as the confirmation
screen displays "<ignored>" for that parameter.