Fix the script to disable teacher theme did not work correctly
The script was not correctly restoring default cursor theme name.
Instead it was setting the cursor theme name to `$default_cursor_theme`.
As the theme was not found, default cursor theme was used instead.
Thus this malfunction got unnoticed.

And as a side result, the log file was not deleted.
This is what led to find the bug.

Now, the default cursor theme name is correctly restored and the log
file is also deleted.