Fix apache was restarted instead of being reloaded
When using the `*website.sh` and `*install-phpmyadmin-from-src.sh`
scripts, apache was displaying a message that it needed to be reloaded.
But the scripts were restarting apache and displaying that it was
restarted.

Now the scripts reload apache and display that apache was reloaded.
Thus it is much more clear for the user that no action needs to be
taken when the script ends.