Run apache as www-data and not as a random user
Previously, apache was run as a random user account choosen during
install time.
But to make the web server multi-user, apache is now run as www-data.

The php fpm pools read and write from and to the sockets using www-data.
But the php fpm pool process is still run as a random user account
choosen durint virtual host and php fpm pool creation.