Fix use of undefined variable `$projects_directory`
The amp install script did not define the `$projects_directory` in the
php fpm teplate file.

Therefore, the script generated an erroneous path for open_basedir path.
Which in turn generated an access denied error when trying to see the
websites' home page.

Now the amp install script does preprocess the php fpm template file and
defines the missing variable.