Fix pma install (from source) fails on ubuntu
Apache is launched as user `www-data`.
So php fpm socket files must be readable by user `www-data`.

This was effective in php fpm pool template file `template-pool.conf`
but not in phpmyadmin php fpm pool file `phpmyadmin-pool.conf`.

This commit sets the user `www-data` as the owner and group of the
php fpm socket file.