Check user id before proceeding
All the scripts must be run as a normal user but need to use sudo.

Now the scripts check that the user is not running the scripts as
root, and then check that the user has the ability to use sudo.

If not, the scripts display an error and stop.