Mkwebsite script creates the virtual host directory if it does not exist
The mkwebsite.sh script did not handle at all the creation of the
virtual host directory.

But when you create a new website, it is useful to create the virtual
host directory automaticaly if it wasn't already there.

Now the script just does that and sets the correct permissions on the
directory.