Remove phpmyadmin config template file from original files directory
The config template file can be found in the phpmyadmin archive file.
So it is very easy to update contrary to other config file which must
be searched in the `/etc` directory of an existing system.

It is better to remove the template file from the `orig` directory and
add a notice in `orig/packages.md` file where the template file can be
found.